Moved sig4 config into package
